requirements.txt
python-dotenv
anthropic
openai
requests
PyYAML
